feat: update landing page and set swap as default
- Relocate the home page to a landing page accessible only manually
- Set the swap route as the default home page route
