+++ src/app/(home)/SwapForm/service/getBalanceNearAllowedToSwap.ts
