+++ src/app/(home)/WarnBox/index.tsx
