+++ src/app/landing/Card/CardInvestorLogo.tsx
