+++ src/app/landing/Card/CardVision.tsx
