+++ src/app/landing/FAQ/index.tsx
