+++ src/app/landing/InvestorLogo/index.tsx
