+++ src/app/landing/PaperHome/index.tsx
