+++ src/components/Modal/ModalReviewSwap.tsx
@@ -5,8 +5,8 @@ import { formatUnits, parseUnits } from "ethers"
 import Image from "next/image"
 import React, { useEffect, useState } from "react"
 
-import { balanceToDecimal } from "@src/app/swap/SwapForm/service/balanceTo"
-import { getBalanceNearAllowedToSwap } from "@src/app/swap/SwapForm/service/getBalanceNearAllowedToSwap"
+import { balanceToDecimal } from "@src/app/(home)/SwapForm/service/balanceTo"
+import { getBalanceNearAllowedToSwap } from "@src/app/(home)/SwapForm/service/getBalanceNearAllowedToSwap"
 import Button from "@src/components/Button/Button"
 import CardSwap from "@src/components/Card/CardSwap"
 import ModalDialog from "@src/components/Modal/ModalDialog"
