+++ src/components/Network/AssetList.tsx
@@ -6,7 +6,7 @@ import React, { type ReactNode } from "react"
 import {
   balanceToCurrency,
   balanceToDecimal,
-} from "@src/app/swap/SwapForm/service/balanceTo"
+} from "@src/app/(home)/SwapForm/service/balanceTo"
 import type { TokenListWithNotSelectableToken } from "@src/components/Modal/ModalSelectAssets"
 import AssetComboIcon from "@src/components/Network/AssetComboIcon"
 import type { NetworkToken } from "@src/types/interfaces"
