+++ src/constants/routes.ts
@@ -1,6 +1,5 @@
 export enum Navigation {
   HOME = "/",
-  SWAP = "/swap",
   DEPOSIT = "/deposit",
   WITHDRAW = "/withdraw",
   WALLET = "/wallet",
@@ -15,7 +14,7 @@ export type NavigationLinks = {
 }
 
 export const LINKS_HEADER: NavigationLinks[] = [
-  { href: Navigation.SWAP, label: "Swap" },
+  { href: Navigation.HOME, label: "Swap" },
   { href: Navigation.DEPOSIT, label: "Deposit" },
   { href: Navigation.WITHDRAW, label: "Withdraw" },
 ]
